:fire: Refactorize 'Object' class name to 'StrictObject'
