+++ src/System/README.MD
@@ -1,1 +1,1 @@
-Las clases de este espacio de nombres no se heredan de NelsonMartell\Object.
+Las clases de este espacio de nombres no se heredan de NelsonMartell\StrictObject.
